pyproject.toml
[tool.poetry]
name = "resume"
version = "0.1.0"
description = "Resume API is a project to create and manage data from a curriculum, which reports a person's educational trajectory and professional experiences, as a way to demonstrate their skills and competences."
authors = ["Wesley Mendes <wesley@wesleymendes.com.br>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
python-decouple = "^3.6"
psycopg2-binary = "^2.9.3"
djangorestframework-simplejwt = "^5.2.0"
django-extensions = "^3.2.0"
PyJWT = "^2.4.0"
db_url = "^0.3.3"
Unipath = "^1.1"
dj-database-url = "^0.5.0"
django-cors-headers = "^3.13.0"
pre-commit = "^2.20.0"
django-admin-sortable2 = "^2.1.8"

[tool.poetry.dev-dependencies]
mypy = "^0.961"
black = "^22.6.0"
factory-boy = "^3.2.0"
Faker = "^9.2.0"
flake8 = "^4.0.1"
pytest = "^7.1.2"
pytest-django = "^4.4.0"
pytest-flake8 = "^1.1.1"
requests = "^2.26.0"
pytest-cov = "^3.0.0"
pytest-pythonpath = "0.7.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true
force-exclude = '''
  (migrations\/.*\.py$|manage\.py|(w|a)sgi\.py|apps\.py|urls\.py|settings\.py)
'''

[tool.isort]
atomic = true
case_sensitive = true
extend_skip = [".json", ".md"]
skip = ["manage.py", "wsgi.py", "asgi.py", "apps.py", "urls.py"]
skip_glob = ["*/migrations/*", "*/tests/*"]
filter_files = true
import_heading_firstparty = "Internal imports"
import_heading_stdlib = "Python imports"
import_heading_thirdparty = "Pip imports"
known_startup = ["startup"]
line_length = 120
lines_after_imports = 2
profile = "black"
py_version = "auto"
remove_redundant_aliases = true
sections = [
  "FUTURE",
  "STARTUP",
  "STDLIB",
  "THIRDPARTY",
  "FIRSTPARTY",
  "LOCALFOLDER",
]
